Initial commit and 1.0.0
With this commit `wtt` has all the functionality I want for a 1.0.0
release. This is essentially a Rust CLI tool replacing these Ruby
scripts:

- https://github.com/sdball/dotfiles/blob/main/bin/insert-time-diff
- https://github.com/sdball/dotfiles/blob/main/bin/inject-time-diff
- https://github.com/sdball/dotfiles/blob/main/bin/inject-time-total

Given JSON lines with a datetime field via filename or STDIN this
tool will either inject new JSON fields or insert lines that track
the time since the previous line and/or time since the first line.